Refactor: エンドポイント周りを整理
- ファイル名/関数名修正
- endpointsパッケージに全てまとめる
