+++ internal/endpoints/doc.go
@@ -1,0 +1,2 @@
+// Package of processes to be initially processed as endpoints.
+package endpoints
