+++ internal/endpoints/welcome.go
@@ -1,11 +1,11 @@
-package welcome
+package endpoints
 
 import (
 	"fmt"
 )
 
 // Welcome screen to be displayed after entering a command.
-func Main() {
+func Welcome() {
 	fmt.Print("    ####       ###             ###      #####      ##         \n")
 	fmt.Print("   ######       ##     ###     ##      ##   ##     ##         \n")
 	fmt.Print("  ##    ##       ##   ## ##   ##       ##          ## ###     \n")
