--- internal/route/main.go
@@ -1,35 +1,0 @@
-package route
-
-import (
-	"awsh/pkg/prompt"
-)
-
-// TODO: GoDocに関数の説明、対象リソースや操作を追加する方法を記載する
-func Main() string {
-	// 操作対象のAWSリソース群
-	route := []string{"S3", "ECS"}
-	var select_action string
-	select_route := prompt.ChooseValueFromPromptItems("Select the resource to be operated", route)
-
-	switch select_route {
-	case "S3":
-		actions := []string{
-			"ListBuckets",
-			"ListObjects",
-			"GetObject",
-			"DownloadObject",
-		}
-
-		select_action = prompt.ChooseValueFromPromptItems("Select an action", actions)
-	case "ECS":
-		actions := []string{
-			"StartECS",
-			"ecs-exec",
-			"StopECSTask",
-		}
-
-		select_action = prompt.ChooseValueFromPromptItems("Select an action", actions)
-	}
-
-	return select_action
-}
